web_app/requirements.txt
appdirs==1.4.3
asgiref==3.2.5
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
cssselect==1.1.0
Django==3.0.4
fake-useragent==0.1.11
gunicorn
idna==2.9
lxml==4.5.0
parse==1.15.0
psycopg2_binary==2.8.4
pyee==7.0.1
pyppeteer==0.0.25
pyquery==1.4.1
pyTelegramBotAPI==3.6.7
pytz==2019.3
requests==2.23.0
requests-html==0.10.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tqdm==4.43.0
urllib3==1.25.8
w3lib==1.21.0
websockets==8.1
Pillow
selenium==3.141.0
